Moved write_all_cores
